Struct sseq::bigraded::DenseBigradedModule

pub struct DenseBigradedModule {
    dimensions: OnceBiVec<OnceBiVec<usize>>,
    min_y: i32,
}

Fields
dimensions: OnceBiVec<OnceBiVec<usize>>min_y: i32
Implementations

impl DenseBigradedModule

pub fn new(min_x: i32, min_y: i32) -> Self

pub const fn min_x(&self) -> i32

pub const fn min_y(&self) -> i32

pub fn max_x(&self) -> i32

pub fn max_y(&self) -> i32

pub fn range(&self, x: i32) -> Range<i32>

pub fn defined(&self, x: i32, y: i32) -> bool

pub fn set_dimension(&self, x: i32, y: i32, dim: usize)

This can only be set when bidegrees to the left and bottom of (x, y) have been set.

pub fn get_dimension(&self, x: i32, y: i32) -> Option<usize>

The dimension in a bidegree, None if not yet defined

pub fn dimension(&self, x: i32, y: i32) -> usize