sseq::sseq
Struct Sseq

pub struct Sseq<P: SseqProfile = Adams> {
    p: ValidPrime,
    classes: Arc<DenseBigradedModule>,
    differentials: BiVec<BiVec<BiVec<Differential>>>,
    permanent_classes: BiVec<BiVec<Subspace>>,
    page_data: BiVec<BiVec<BiVec<Subquotient>>>,
    invalid: BiVec<BiVec<bool>>,
    profile: PhantomData<P>,
}

Fields
p: ValidPrimeclasses: Arc<DenseBigradedModule>

The first page of the spectral sequence

differentials: BiVec<BiVec<BiVec<Differential>>>

x -> y -> r -> differential

If the bidegree is valid (see Sseq::invalid), then the differential is reduced.

permanent_classes: BiVec<BiVec<Subspace>>

x -> y -> permanent_classes

page_data: BiVec<BiVec<BiVec<Subquotient>>>

x -> y -> r -> E_r^{x, y} as a subquotient of the original bidegree.

The “quotient” part of the subquotient is always accurate, but the “sub” part may not be. The invalid field tracks which bidegrees are valid.

This is always the full ambient space when r = P::MIN_R, but we set min_degree to P::MIN_R to make code a bit more streamlined.

Invariants:

if differential[x][y][r] is defined, then page_data[x][y][r + 1] and page_data[tx][ty][r + 1] are always defined,

invalid: BiVec<BiVec<bool>>

x -> y -> validity. A bidegree is invalid if the page_data is no longer accurate.

profile: PhantomData<P>
Implementations

impl<P: SseqProfile> Sseq<P>

pub fn new(p: ValidPrime, min_x: i32, min_y: i32) -> Self

pub fn min_x(&self) -> i32

pub fn min_y(&self) -> i32

pub fn classes(&self) -> Arc<DenseBigradedModule>

pub fn range(&self, x: i32) -> Range<i32>

pub fn max_x(&self) -> i32

pub fn max_y(&self) -> i32

pub fn defined(&self, x: i32, y: i32) -> bool

pub fn set_dimension(&mut self, x: i32, y: i32, dim: usize)

pub fn clear(&mut self)

pub fn dimension(&self, x: i32, y: i32) -> usize

pub fn add_permanent_class( &mut self, x: i32, y: i32, class: FpSlice<'_>, ) -> bool

Returns

Whether a new permanent class was added

fn extend_differential(&mut self, r: i32, x: i32, y: i32)

Ensure self.differentials[x][y][r] is defined. Must call extend_page_data on the source and target after this.

fn extend_page_data(&mut self, r: i32, x: i32, y: i32)

Ensure self.page_data[x][y][r] is defined

pub fn add_differential( &mut self, r: i32, x: i32, y: i32, source: FpSlice<'_>, target: FpSlice<'_>, ) -> bool

Add a $d_r$ differential from bidegree $(x, y)$, with the given source and target classes.

Return

Whether the differential is new

pub fn invalid(&self, x: i32, y: i32) -> bool

pub fn update(&mut self)

pub fn update_bidegree(&mut self, x: i32, y: i32) -> BiVec<Vec<Vec<u32>>>

This returns the vec of differentials to draw on each page.

pub fn complete(&self, x: i32, y: i32) -> bool

Whether the calcuations at bidegree (x, y) are complete. This means all classes on the final page are known to be permanent.

pub fn inconsistent(&self, x: i32, y: i32) -> bool

Whether there is an inconsistent differential involving bidegree (x, y).

pub fn differentials(&self, x: i32, y: i32) -> &BiVec<Differential>

pub fn differentials_hitting( &self, x: i32, y: i32, ) -> impl Iterator<Item = (i32, &Differential)> + '_

pub fn permanent_classes(&self, x: i32, y: i32) -> &Subspace

pub fn page_data(&self, x: i32, y: i32) -> &BiVec<Subquotient>

pub fn multiply( &self, x: i32, y: i32, class: FpSlice<'_>, prod: &Product, ) -> Option<(i32, i32, FpVector)>

Compute the product between product and the class class at (x, y). Returns None if the product is not yet computed.

pub fn leibniz( &mut self, r: i32, x: i32, y: i32, class: FpSlice<'_>, source_product: &Product, target_product: Option<&Product>, ) -> Option<(i32, i32, i32, FpVector)>

Apply the Leibniz rule to obtain new differentials. The differential we start with is a d_r differential from (x, y) with source s and target t. If the source is permanent, then r should be set to i32::MAX.

Arguments

source_product the product to multiply the class with

target_product the differential on source_product. If source_product is permanent, then this is None.

Return

We return a tuple (r, x, y, class) recording the (source of) the new differential. If the function returns None, this means no differential was added. This can either be because the differential was trivial, or the data needed to compute the product is not yet available.

pub fn write_to_graph<'a, T: Backend>( &self, g: T, r: i32, differentials: bool, products: impl Iterator<Item = &'a (String, Product)> + Clone, header: impl FnOnce(&mut T) -> Result<(), T::Error>, ) -> Result<(), T::Error>

This shifts the sseq horizontally so that the minimum x is 0.