Struct sseq::Differential

pub struct Differential {
    pub matrix: Matrix,
    first_empty_row: usize,
    source_dim: usize,
    target_dim: usize,
    error: bool,
}

Fields
matrix: Matrixfirst_empty_row: usizesource_dim: usizetarget_dim: usizeerror: bool
Implementations

impl Differential

pub fn new(p: ValidPrime, source_dim: usize, target_dim: usize) -> Self

pub fn set_to_zero(&mut self)

pub fn add(&mut self, source: Slice<'_>, target: Option<Slice<'_>>) -> bool

Add a differential

Return

Whether a new differential was indeed added. If false, the differential is already existent.

Arguments

source: The source of the differential

target: The target of the differential. If None, the differential is zero. This should be reduced by the known images of earlier differentials.

pub fn get_source_target_pairs(&self) -> Vec<(FpVector, FpVector)>

An iterator of differentials in the form (source, target)

pub fn reduce_target(&mut self, zeros: &Subspace)

Given a subspace of the target space, project the target vectors to the complement.

pub fn evaluate(&self, source: Slice<'_>, target: SliceMut<'_>)

This evaluates the differential on source, adding the result to target. This assumes all unspecified differentials are zero. More precisely, it assumes every non-pivot column of the differential matrix has zero differential. This may or may not be actually true (e.g. if we only know d(a + b) = c, it might be that d(a) = c and d(b) = 0, or vice versa, or neither. Here we assume d(a) = c and d(b) = 0.

pub fn prime(&self) -> ValidPrime

pub fn inconsistent(&self) -> bool

Whether the current set of differentials is inconsistent. This should be called only after reduce_target is called.